Lockdown actions on org settings route

+++ portal/src/routes/(authenticated)/organizations/[id]/settings/groups/+page.server.ts
@@ -22,12 +22,23 @@ export const actions = {
   async addGroup(event) {
     const form = await superValidate(event.request, valibot(addGroupSchema));
     if (!form.valid) return fail(400, { form, ok: false });
-    await DatabaseWrites.groups.createGroup(form.data.name, form.data.abbreviation, form.data.orgId);
+    await DatabaseWrites.groups.createGroup(
+      form.data.name,
+      form.data.abbreviation,
+      form.data.orgId
+    );
     return { form, ok: true };
   },
   async deleteGroup(event) {
     const form = await superValidate(event.request, valibot(deleteSchema));
     if (!form.valid) return fail(400, { form, ok: false });
+    if ( // if user modified hidden values
+      !(await prisma.groups.findFirst({
+        where: { Id: form.data.id, OwnerId: parseInt(event.params.id) }
+      }))
+    ) {
+      return fail(403, { form, ok: false });
+    }
     return { form, ok: await DatabaseWrites.groups.deleteGroup(form.data.id) };
   }
 } satisfies Actions;
